20121130
* Not compatible with Android 4.1.x!

* All APKs updated to 4.2.0 versions, where applicable.

* ThinkFree and Microbes live wallpapers removed to reduce
size.

* Still allows sync between AOSP Calendar.apk and Google
Calendar.

* Does not include any non-NEON libraries, as there have not
been any non-NEON devices (Tegra and Tegra 2 devices such as
the Motorola XOOM) which have received 4.2, and the 4.1
libraries are not compatible with 4.2.

* Installs Photo Sphere and Picasa sync on Mako (Nexus 4),
Manta (Nexus 10), and Tuna (Galaxy Nexus, all variants).

** Does not install on Grouper (Nexus 7) because it does not
have a rear camera.

** The GalleryGoogle.apk for each is different, thus why there
is a folder and APK for each device.

** This will override any custom ROM camera features on all
three, since it replaces the ROM's camera/gallery APK.

* system/etc/g.prop updated:

** Users on 4.1.x will not receive GooManager update
notifications for this and future 4.2.x packages.

** Users on any past or third-party GApps packages and 4.2.x
will also not be recommended this package. This is because the
platform version listed is now jb42 instead of jb.